invoice_process/requirements.txt
langchain
langchain-community
langchain-core
langchain-ollama
pydantic
python-dotenv
wikipedia
mcp
langgraph
ollama
fastmcp
fastapi
uvicorn
ipython
langgraph-prebuilt
langsmith
mermaid-py
langchain-tavily
langgraph-checkpoint-sqlite
pdf2image